justsmile/requirements.txt
django
djangorestframework
overpy